Some API fixes, and tag picker based on role fixes

+++ app/Http/Controllers/Api/ArchiveApi/Commands.php
@@ -5,10 +5,12 @@ namespace App\Http\Controllers\Api\ArchiveApi;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Illuminate\Support\Facades\Validator;
+
 use App\Helpers\Generator;
+use App\Helpers\Validation;
 
 use App\Models\Archive;
+use App\Models\History;
 use App\Models\ArchiveRelation;
 
 class Commands extends Controller
@@ -16,37 +18,62 @@ class Commands extends Controller
     //
     public function createArchive(Request $request)
     {
-        //Helpers
-        $slug = Generator::getSlugName($request->archive_name, "archive");
-
         try{
-            $validator = Validator::make($request->all(), [
-                'archive_name' => 'required',
-                'archive_desc' => 'required'
-            ]);
+            $slug = Generator::getSlugName($request->archive_name, "archive");
+            $validator = Validation::getValidateArchive($request);
+            $user_id = $request->user()->id;
 
             if ($validator->fails()) {
-                return response()->json([
-                    'status' => 'error',
-                    'message' => $validator->errors()
-                ], Response::HTTP_BAD_REQUEST);
-            } else {
-                $archive = Archive::create([
-                    'id' => Generator::getUUID(),
-                    'slug_name' => $slug,
-                    'archive_name' => $request->archive_name,
-                    'archive_desc' => $request->archive_desc,
-                    'created_by' => 1, // for now, later will be changed to auth user
-                    'created_at' => date('Y-m-d H:i:s'),
-                    'updated_by' => null,
-                    'updated_at' => null
-                ]);
+                $errors = $validator->messages();
 
                 return response()->json([
-                    'status' => 'success',
-                    'message' => 'Archive Created',
-                    'data' => $archive
-                ], Response::HTTP_OK);
+                    'status' => 'failed',
+                    'result' => $errors,
+                ], Response::HTTP_UNPROCESSABLE_ENTITY);
+            } else {
+                $data = new Request();
+                $obj = [
+                    'history_type' => "archive",
+                    'history_body' => "has created a archive called '".$request->archive_name."'"
+                ];
+                $data->merge($obj);
+
+                $validatorHistory = Validation::getValidateHistory($data);
+                if ($validatorHistory->fails()) {
+                    $errors = $validatorHistory->messages();
+
+                    return response()->json([
+                        'status' => 'failed',
+                        'result' => $errors,
+                    ], Response::HTTP_UNPROCESSABLE_ENTITY);
+                } else {
+                    $archive = Archive::create([
+                        'id' => Generator::getUUID(),
+                        'slug_name' => $slug,
+                        'archive_name' => $request->archive_name,
+                        'archive_desc' => $request->archive_desc,
+                        'created_by' => $user_id,
+                        'created_at' => date('Y-m-d H:i:s'),
+                        'updated_by' => null,
+                        'updated_at' => null
+                    ]);
+
+                    History::create([
+                        'id' => Generator::getUUID(),
+                        'history_type' => $data->history_type, 
+                        'context_id' => $archive->id, 
+                        'history_body' => $data->history_body, 
+                        'history_send_to' => null,
+                        'created_at' => date("Y-m-d h:i:s"),
+                        'created_by' => $user_id
+                    ]);
+
+                    return response()->json([
+                        'status' => 'success',
+                        'message' => 'Archive Created',
+                        'data' => $archive
+                    ], Response::HTTP_OK);
+                }
             }
         } catch(\Exception $e) {
             return response()->json([
@@ -57,11 +84,10 @@ class Commands extends Controller
     }
 
     public function addToArchive(Request $request){
-
+        // This usecase doens't need to include history !
         try{
-
             $user_id = $request->user()->id;
-
+            
             $relation = ArchiveRelation::create([
                 'id' => Generator::getUUID(),
                 'archive_id' => $request->archive_id,
@@ -86,26 +112,68 @@ class Commands extends Controller
     public function editArchive(Request $request, $id)
     {
         try{
-            //Validate name avaiability
-            $check = Archive::where('archive_name', $request->archive_name)->where('id_user', $request->id_user)->get();
+            $user_id = $request->user()->id;
+           
+            $check = Archive::where('archive_name', $request->archive_name)
+                ->where('id_user', $user_id)->get();
 
             if(count($check) == 0){
-                $result = Archive::where('id', $id)->update([
-                    'archive_name' => $request->archive_name,
-                    'updated_at' => date("Y-m-d h:i")
-                ]);
+                $validator = Validation::getValidateArchive($request);
 
-                return response()->json([
-                    'status' => 'success',
-                    'message' => 'Archive successfully updated',
-                    'data' => $result
-                ], Response::HTTP_OK);
+                if ($validator->fails()) {
+                    $errors = $validator->messages();
+    
+                    return response()->json([
+                        'status' => 'failed',
+                        'result' => $errors
+                    ], Response::HTTP_UNPROCESSABLE_ENTITY);
+                } else {
+                    $data = new Request();
+                    $obj = [
+                        'history_type' => "archive",
+                        'history_body' => "has updated '".$request->archive_name."' archive"
+                    ];
+                    $data->merge($obj);
+    
+                    $validatorHistory = Validation::getValidateHistory($data);
+                    if ($validatorHistory->fails()) {
+                        $errors = $validatorHistory->messages();
+    
+                        return response()->json([
+                            'status' => 'failed',
+                            'result' => $errors,
+                        ], Response::HTTP_UNPROCESSABLE_ENTITY);
+                    } else {
+                        $result = Archive::where('id', $id)->update([
+                            'archive_name' => $request->archive_name,
+                            'archive_desc' => $request->archive_desc,
+                            'updated_at' => date("Y-m-d h:i"),
+                            'updated_by' => $user_id
+                        ]);
+
+                        History::create([
+                            'id' => Generator::getUUID(),
+                            'history_type' => $data->history_type, 
+                            'context_id' => $id, 
+                            'history_body' => $data->history_body, 
+                            'history_send_to' => null,
+                            'created_at' => date("Y-m-d h:i:s"),
+                            'created_by' => $user_id
+                        ]);
+
+                        return response()->json([
+                            'status' => 'success',
+                            'message' => 'Archive successfully updated',
+                            'data' => $result
+                        ], Response::HTTP_OK);
+                    }
+                }
             } else {
                 return response()->json([
                     'status' => 'failed',
                     'message' => 'Archive name must be unique',
                     'data' => null
-                ], Response::HTTP_OK);
+                ], Response::HTTP_UNPROCESSABLE_ENTITY);
             }
         } catch (\Exception $e) {
             return response()->json([
@@ -118,20 +186,45 @@ class Commands extends Controller
     public function deleteArchive(Request $request, $id)
     {
         try {
-
             $user_id = $request->user()->id;
 
-            $result = Archive::destroy($id);
+            $data = new Request();
+            $obj = [
+                'history_type' => "archive",
+                'history_body' => "has deleted '".$request->archive_name."' archive"
+            ];
+            $data->merge($obj);
 
-            //Delete archive relation
-            ArchiveRelation::where('archive_id', $id)
-                ->where('user_id', $user_id)
-                ->delete();
+            $validatorHistory = Validation::getValidateHistory($data);
+            if ($validatorHistory->fails()) {
+                $errors = $validatorHistory->messages();
 
-            return response()->json([
-                'status' => 'success',
-                'message' => 'Archive deleted'
-            ], Response::HTTP_OK);
+                return response()->json([
+                    'status' => 'failed',
+                    'result' => $errors,
+                ], Response::HTTP_UNPROCESSABLE_ENTITY);
+            } else {
+                $result = Archive::destroy($id);
+
+                ArchiveRelation::where('archive_id', $id)
+                    ->where('user_id', $user_id)
+                    ->delete();
+
+                History::create([
+                    'id' => Generator::getUUID(),
+                    'history_type' => $data->history_type, 
+                    'context_id' => null, 
+                    'history_body' => $data->history_body, 
+                    'history_send_to' => null,
+                    'created_at' => date("Y-m-d h:i:s"),
+                    'created_by' => $user_id
+                ]);
+
+                return response()->json([
+                    'status' => 'success',
+                    'message' => 'Archive deleted'
+                ], Response::HTTP_OK);
+            }
         } catch (\Exception $e) {
             return response()->json([
                 'status' => 'error',
